backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
snowflake-snowpark-python>=1.11.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
redis>=5.0.0
tenacity>=8.2.3
structlog>=23.2.0
typing-extensions>=4.5.0

